타이머 관련 오류수정 및 전체화면 이미지 업데이트 로직

+++ SevenDays 새버전/sevenDays/ViewController/Data/DataModel.swift
@@ -21,8 +21,8 @@ enum Language {
 // 프라퍼티 설명:  언어 설정, 배경음, 효과음, 넘김 속도
 struct Setting {
     var langueage: Language
-    var bgmVolume: Double
-    var effectVolume: Double
+    var bgmVolume: Float
+    var effectVolume: Float
     var textSpeed: Double
     
 }
